moving the RAG tools to a new file
